change box shadow

+++ Finanzas.sql
@@ -18,10 +18,10 @@ CREATE TABLE Finanzas(
     id INT AUTO_INCREMENT,
     fk_id_usuario INT,
     mes VARCHAR(100),
-    ingreso INT,
+    ingreso FLOAT,
     ahorroEsperado VARCHAR(100),
-    gastoTotal INT,
-    ahorroReal INT,
+    gastoTotal FLOAT,
+    ahorroReal FLOAT,
     PRIMARY KEY (id),
     FOREIGN KEY(fk_id_usuario) REFERENCES Usuarios(id) ON UPDATE CASCADE ON DELETE CASCADE
 );
@@ -32,14 +32,6 @@ select * from Finanzas where fk_id_usuario = 1;
 
 SELECT * FROM Finanzas;
 
-#CREATE TABLE Usuarios_Finanzas(
-#	id INT AUTO_INCREMENT,
-#	fk_id_usuario INT NOT NULL,
-#	fk_id_finanzas INT NOT NULL,
-#	PRIMARY KEY (id),
-#	FOREIGN KEY(fk_id_usuario) REFERENCES Usuarios(id) ON UPDATE CASCADE ON DELETE CASCADE, 
-#	FOREIGN KEY(fk_id_finanzas) REFERENCES Finanzas(id) ON UPDATE CASCADE ON DELETE CASCADE
-#);
 
 #DROP TABLE Gastos;
 
@@ -50,8 +42,8 @@ CREATE TABLE Gastos (
     tipo VARCHAR(100),
     dia INT,
     descripcion VARCHAR(100),
-    precio INT,
-    total INT,
+    precio FLOAT,
+    total FLOAT,
     PRIMARY KEY (id)
 );
